+++ packages/esm-patient-procedures-app/src/workspaces/procedures-form/procedures-form.scss
