+++ packages/esm-patient-procedures-app/translations/en.json
@@ -3,31 +3,72 @@
   "addProcedure": "Add procedure",
   "am": "AM",
   "amPm": "AM/PM",
+  "april": "April",
+  "august": "August",
   "bodySite": "Body site",
   "cancel": "Cancel",
   "date": "Date",
+  "december": "December",
   "delete": "Delete",
   "deleteModalConfirmationText": "Are you sure you want to delete this procedure?",
   "deleteProcedure": "Delete procedure",
   "deleting": "Deleting",
   "duration": "Duration",
+  "durationUnit": "Duration unit",
+  "durationValue": "Duration",
   "edit": "Edit",
   "editProcedure": "Edit procedure",
   "endDate": "End date",
+  "endDateAndTime": "End date and time",
+  "enterBodySite": "Enter body site",
+  "enterDuration": "Enter duration",
+  "enterNotes": "Enter notes (optional)",
+  "enterProcedure": "Enter procedure",
   "errorDeletingProcedure": "Error deleting procedure",
+  "errorSavingProcedure": "Error saving procedure",
+  "estimatedStartDate": "Estimated start date",
+  "february": "February",
+  "isStartDateKnown": "Is start date known?",
+  "january": "January",
+  "july": "July",
+  "june": "June",
+  "loading": "Loading",
+  "march": "March",
+  "may": "May",
+  "monthOptional": "Month (optional)",
+  "no": "No",
   "noProceduresToDisplay": "No procedures to display",
   "noResultsFor": "No results for",
   "notes": "Notes",
+  "november": "November",
+  "october": "October",
   "pm": "PM",
   "procedure": "Procedure",
   "procedureDeleted": "Procedure deleted",
+  "procedureDuration": "Procedure duration",
+  "procedureNowVisible": "It is now visible on the Procedures page",
   "procedures": "Procedures",
   "procedures_lower": "procedures",
+  "procedureSaved": "Procedure saved",
   "procedureType": "Procedure type",
   "recordProcedure": "Record procedure",
+  "required": "Required",
+  "saveAndClose": "Save & close",
+  "saving": "Saving",
+  "searchBodySites": "Search body sites",
   "searching": "Searching",
+  "searchProcedures": "Search procedures",
   "seeAll": "See all",
+  "selectDurationUnit": "Select unit",
+  "selectMonth": "Select month (optional)",
+  "selectProcedureType": "Select procedure type",
+  "selectStatus": "Select status",
+  "selectYear": "Select year",
+  "september": "September",
   "startDate": "Start date",
+  "startDateAndTime": "Start date and time",
   "status": "Status",
-  "time": "Time"
+  "time": "Time",
+  "year": "Year",
+  "yes": "Yes"
 }
